#302 - refactor class names
