+++ src/main/java/com/commercetools/sync/producttypes/utils/PlainEnumsUpdateActionUtils.java
@@ -8,18 +8,18 @@ import io.sphere.sdk.producttypes.commands.updateactions.ChangeEnumValueOrder;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.List;
 import java.util.Collections;
+import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdateEnumActionsUtils.buildRemoveEnumValuesUpdateActions;
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdateEnumActionsUtils.buildMatchingEnumValuesUpdateActions;
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdateEnumActionsUtils.buildAddEnumValuesUpdateActions;
-import static com.commercetools.sync.producttypes.utils.ProductTypeUpdateEnumActionsUtils.buildChangeEnumValuesOrderUpdateAction;
+import static com.commercetools.sync.producttypes.utils.EnumUpdateActionsUtils.buildAddEnumValuesUpdateActions;
+import static com.commercetools.sync.producttypes.utils.EnumUpdateActionsUtils.buildChangeEnumValuesOrderUpdateAction;
+import static com.commercetools.sync.producttypes.utils.EnumUpdateActionsUtils.buildMatchingEnumValuesUpdateActions;
+import static com.commercetools.sync.producttypes.utils.EnumUpdateActionsUtils.buildRemoveEnumValuesUpdateActions;
 import static java.util.Collections.emptyList;
 
-public final class ProductTypeUpdatePlainEnumActionUtils {
+public final class PlainEnumsUpdateActionUtils {
     /**
      * Compares a list of old {@link EnumValue}s with a list of new {@link EnumValue}s for a given
      * attribute definition.
@@ -84,7 +84,7 @@ public final class ProductTypeUpdatePlainEnumActionUtils {
                 attributeDefinitionName,
                 oldEnumValues,
                 newEnumValues,
-                PlainEnumUpdateActionsUtils::buildActions
+                PlainEnumUpdateActionUtils::buildActions
             );
 
         final List<UpdateAction<ProductType>> addEnumValuesUpdateActions = buildAddEnumValuesUpdateActions(
@@ -118,6 +118,6 @@ public final class ProductTypeUpdatePlainEnumActionUtils {
     }
 
 
-    private ProductTypeUpdatePlainEnumActionUtils() {
+    private PlainEnumsUpdateActionUtils() {
     }
 }
