+++ src/test/java/com/commercetools/sync/producttypes/utils/LocalizedEnumValueUpdateActionUtilsTest.java
@@ -9,7 +9,7 @@ import org.junit.Test;
 
 import java.util.Optional;
 
-import static com.commercetools.sync.producttypes.utils.LocalizedEnumUpdateActionsUtils.buildChangeLabelAction;
+import static com.commercetools.sync.producttypes.utils.LocalizedEnumUpdateActionUtils.buildChangeLabelAction;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class LocalizedEnumValueUpdateActionUtilsTest {
