+++ src/test/java/com/commercetools/sync/producttypes/utils/PlainEnumValueUpdateActionUtilsTest.java
@@ -9,7 +9,7 @@ import org.junit.Test;
 
 import java.util.Optional;
 
-import static com.commercetools.sync.producttypes.utils.PlainEnumUpdateActionsUtils.buildChangeLabelAction;
+import static com.commercetools.sync.producttypes.utils.PlainEnumUpdateActionUtils.buildChangeLabelAction;
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class PlainEnumValueUpdateActionUtilsTest {
